% Generated by roxygen2 (4.0.2): do not edit by hand
\name{conSubseq}
\alias{conSubseq}
\title{W1}
\usage{
conSubseq(x, y, k, ysign = "both")
}
\arguments{
\item{x}{sequence 1}

\item{y}{sequence 2}

\item{k}{length of subsequences}

\item{ysign}{indicating whether only the positive indicators ("pos"), only the negative indicators ("neg"), or both ("both") should be counted}
}
\description{
A wrapper function that calculates W1 and its variants, calling \code{conSubseq_small_k} for \code{k} less than 100 and \code{conSubseq_large_k} for larger \code{k}.
}

