% Generated by roxygen2 (4.0.2): do not edit by hand
\name{conSubseq_small_k}
\alias{conSubseq_small_k}
\title{W1 for small k}
\usage{
conSubseq_small_k(x, y, k, ysign = "both")
}
\arguments{
\item{x}{sequence 1}

\item{y}{sequence 2}

\item{k}{length of subsequences}

\item{ysign}{indicating whether only the positive indicators ("pos"), only the negative indicators ("neg"), or both ("both") should be counted}
}
\description{
Calculate W1 and its variants. The algorithm has a running time of O(k^2 n) and is recommended for small \code{k}.
}

