% Generated by roxygen2 (4.0.2): do not edit by hand
\name{conSubseq_varEst}
\alias{conSubseq_varEst}
\title{Variance estimation for W1}
\usage{
conSubseq_varEst(n, k, m, ysign = "both")
}
\arguments{
\item{n}{length of sequences}

\item{k}{length of subsequences}

\item{m}{number of random sequence pairs to generate}

\item{ysign}{indicating whether only the positive indicators ("pos"), only the negative indicators ("neg"), or both ("both") should be counted}
}
\description{
Estimate variance for the W1 measure using Monte Carlo simulations on randomly generated sequence pairs that are independent of each other.
}

