% Generated by roxygen2 (4.0.2): do not edit by hand
\name{incSubseq}
\alias{incSubseq}
\title{W2}
\usage{
incSubseq(x, y, k, ysign = "both")
}
\arguments{
\item{x}{sequence 1}

\item{y}{sequence 2}

\item{k}{length of subsequences}

\item{ysign}{indicating whether only the positive indicators ("pos"), only the negative indicators ("neg"), or both ("both") should be counted}
}
\description{
A wrapper function that calculates W2 and its variants, calling \code{incSubseq_small_n} for \code{n} less than 5000 and \code{incSubseq_large_n} for larger \code{n}.
}

