% Generated by roxygen2 (4.0.2): do not edit by hand
\name{incSubseq_large_n}
\alias{incSubseq_large_n}
\title{W2 for large n}
\usage{
incSubseq_large_n(x, y, k, ysign = "both")
}
\arguments{
\item{x}{sequence 1}

\item{y}{sequence 2}

\item{k}{length of subsequences}

\item{ysign}{indicating whether only the positive indicators ("pos"), only the negative indicators ("neg"), or both ("both") should be counted}
}
\description{
Calculate W2 and its variants. The algorithm has a running time of O(knlog(n)), but runs slower than \code{incSubseq_small_n} for small \code{n}; recommended for large \code{n}.
}

