% Generated by roxygen2 (4.0.2): do not edit by hand
\name{incSubseq_pval}
\alias{incSubseq_pval}
\title{Asymptotic p-value for W2}
\usage{
incSubseq_pval(x, y, k, m, ysign = "both")
}
\arguments{
\item{x}{sequence 1}

\item{y}{sequence 2}

\item{k}{length of subsequences}

\item{m}{number of random sequence pairs to generate}

\item{ysign}{indicating whether only the positive indicators ("pos"), only the negative indicators ("neg"), or both ("both") should be counted}
}
\value{
\code{coexp_meas} value of W2

\code{est_var} estimated variance of W2

\code{coexp_meas_norm} value of T2

\code{p_val} asymptotic p-value
}
\description{
An overall function that calculates W2 using \code{incSubseq}, estimates the variance using \code{incSubseq_varEst}, computes the normalized statistic T2 and the asymptotic p-value. Returns all four outputs.
}

